requirements.txt
flask==2.0.1
spotipy==2.19.0
pandas==1.4.2
scikit-learn==1.0.2
numpy==1.22.3
python-dotenv==0.19.2
requests==2.27.1 